# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Studio Guest Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /known-guests:
    get:
      operationId: listKnownGuests
      tags: [known-guests]
      summary: List known (returning) guests with visit stats
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: vip
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Known guests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnownGuest"
  /known-guests/{id}:
    patch:
      operationId: updateKnownGuest
      tags: [known-guests]
      summary: Update a known guest (VIP flag)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnownGuestUpdate"
      responses:
        "200":
          description: Updated known guest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnownGuest"
        "404":
          description: Not found
  /known-guests/{id}/visits:
    get:
      operationId: listKnownGuestVisits
      tags: [known-guests]
      summary: Visit history for a known guest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Past visits (guest records)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Guest"
        "404":
          description: Not found
  /kiosk/preregistrations:
    get:
      operationId: kioskListPreregistrations
      tags: [kiosk]
      summary: Search today's pending pre-registrations by guest name (kiosk self check-in)
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
      responses:
        "200":
          description: Matching expected guests (limited fields)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KioskPreregistration"
        "400":
          description: Invalid query
  /kiosk/checkin:
    post:
      operationId: kioskCheckin
      tags: [kiosk]
      summary: Self check-in for an expected guest (converts the pre-registration)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KioskCheckinRequest"
      responses:
        "201":
          description: Guest checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
        "404":
          description: Pre-registration not found or no longer pending
        "403":
          description: Entry requires assistance at the security desk
...
components:
  schemas:
...
    Guest:
      type: object
      required:
        - id
        - name
        - company
        - hostName
        - purposeOfVisit
        - site
        - status
        - checkinAt
        - badgeId
      properties:
        id:
          type: integer
        badgeId:
          type: string
        name:
          type: string
        company:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        hostName:
          type: string
        purposeOfVisit:
          type: string
        site:
          type: string
          description: "Dallas/The Plex, Tustin, Nashville"
        status:
          type: string
          enum: [active, checked_out]
        checkinAt:
          type: string
          format: date-time
        checkoutAt:
          type: ["string", "null"]
          format: date-time
        expectedDeparture:
          type: ["string", "null"]
          format: date-time
        photoUrl:
          type: ["string", "null"]
        checkedInByClerkId:
          type: ["string", "null"]
        checkedOutByClerkId:
          type: ["string", "null"]
        preregistrationId:
          type: ["integer", "null"]
        studios:
          type: array
          items:
            type: string
        timeOnSiteMinutes:
          type: ["integer", "null"]
        isOverdue:
          type: boolean
...
    KnownGuest:
      type: object
      required:
        - id
        - name
        - isVip
        - visitCount
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        company:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        isVip:
          type: boolean
        visitCount:
          type: integer
        lastVisitAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    KnownGuestUpdate:
      type: object
      required:
        - isVip
      properties:
        isVip:
          type: boolean
...
    KioskPreregistration:
      type: object
      required:
        - id
        - guestName
        - company
        - hostName
        - expectedArrival
      properties:
        id:
          type: integer
        guestName:
          type: string
        company:
          type: string
        hostName:
          type: string
        expectedArrival:
          type: string
          format: date-time
    KioskCheckinRequest:
      type: object
      required:
        - preregistrationId
      properties:
        preregistrationId:
          type: integer
        photoUrl:
          type: ["string", "null"]
...