# lib/api-zod/src/generated/types/kioskCheckinRequest.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Studio Guest Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface KioskCheckinRequest {
  preregistrationId: number;
  /** @nullable */
  photoUrl?: string | null;
}