# lib/api-zod/src/generated/types/kioskPreregistration.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Studio Guest Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface KioskPreregistration {
  id: number;
  guestName: string;
  company: string;
  hostName: string;
  expectedArrival: Date;
}